Feature/integrate user app

+++ app/src/main/java/com/future/tailormade/core/model/response/history/OrderDetailMeasurementResponse.kt
@@ -1,9 +1,13 @@
 package com.future.tailormade.core.model.response.history
 
 data class OrderDetailMeasurementResponse(
-    val chest: Int,
-    val hips: Int,
-    val inseam: Int,
-    val neckToWaist: Int,
-    val waist: Int
-)
+
+    val chest: Float,
+
+    val hips: Float,
+
+    val inseam: Float,
+
+    val neckToWaist: Float,
+
+    val waist: Float)
